// src/lib/api/types.ts
export type SponsorBlockCategory =
  | 'sponsor' | 'selfpromo' | 'interaction' | 'poi_highlight' | 'intro'
  | 'outro' | 'preview' | 'hook' | 'filler' | 'music_offtopic'

export type SponsorBlockOption = 'auto_skip' | 'manual_skip' | 'show_in_seekbar' | 'disable'

export interface SponsorBlockSegment {
  category: SponsorBlockCategory
  actionType: 'skip' | 'mute' | 'full' | 'poi'
  segment: [number, number]
  UUID: string
  videoDuration: number
  locked: number
  votes: number
  description: string
}